+++ EscapeOS/Platform/TabBarConfigurator.h
@@ -1,4 +1,0 @@
-#import <UIKit/UIKit.h>
-
-/// Applies system tab bar styling. On iOS 26+, also enables minimize-on-scroll when supported at runtime.
-void EscapeOSConfigureTabBarController(UITabBarController *controller);